refactor: Answer 패키지 삭제
창하의 동시성 인수테스트는.. 아직 공개되지 않은 청원 게시물에 대해서 답변을 다는 시도를 하기 때문에 에러가 발생함.

지금까지 우리 서비스는 어떠한 공개가 되지 않은 글에도 답변을 달 수 있는 구조였다는 것..
