--- src/main/java/com/gistpetition/api/answer/dto/AnswerRequest.java
@@ -1,20 +1,0 @@
-package com.gistpetition.api.answer.dto;
-
-import javax.validation.constraints.NotBlank;
-
-public class AnswerRequest {
-
-    @NotBlank
-    private String content;
-
-    protected AnswerRequest() {
-    }
-
-    public AnswerRequest(String content) {
-        this.content = content;
-    }
-
-    public String getContent() {
-        return content;
-    }
-}
