+++ src/main/java/com/gistpetition/api/petition/domain/Answer.java
@@ -1,12 +1,13 @@
-package com.gistpetition.api.answer.domain;
+package com.gistpetition.api.petition.domain;
 
 import com.gistpetition.api.common.persistence.BaseEntity;
 import lombok.Getter;
 import org.hibernate.envers.Audited;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.JoinColumn;
 import javax.persistence.Lob;
+import javax.persistence.OneToOne;
 
 @Audited
 @Entity
@@ -14,18 +15,23 @@ import javax.persistence.Lob;
 public class Answer extends BaseEntity {
     @Lob
     private String content;
-    @Column(unique = true)
-    private Long petitionId;
+    @OneToOne
+    @JoinColumn(name = "petition_id", referencedColumnName = "id", unique = true)
+    private Petition petition;
 
     protected Answer() {
     }
 
-    public Answer(String content, Long petitionId) {
+    public Answer(String content, Petition petition) {
         this.content = content;
-        this.petitionId = petitionId;
+        this.petition = petition;
     }
 
     public void updateContent(String content) {
         this.content = content;
+    }
+
+    public void detach() {
+        this.petition = null;
     }
 }
