+++ src/main/java/com/gistpetition/api/petition/domain/AnswerRepository.java
@@ -1,0 +1,11 @@
+package com.gistpetition.api.petition.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface AnswerRepository extends JpaRepository<Answer, Long> {
+    List<Answer> findAllByPetitionId(Long petitionId);
+}
